sync with upstream

+++ addons/sourcemod/scripting/freak_fortress_2/bosses.sp
@@ -740,6 +740,7 @@ static void LoadCharacter(const char[] character, int charset, const char[] map,
 	{
 		int special = BossList.Length;
 		bool clean = !Cvar[Debugging].BoolValue;
+		bool check = Cvar[FileCheck].BoolValue;
 		
 		snap = cfg.Snapshot();
 		
@@ -791,7 +792,7 @@ static void LoadCharacter(const char[] character, int charset, const char[] map,
 					{
 						if(cfgsub.Get("worldmodel", buffer, sizeof(buffer)) && buffer[0])
 						{
-							if(FileExists(buffer, true))
+							if(!check || FileExists(buffer, true))
 							{
 								cfgsub.SetInt("worldmodel", PrecacheModel(buffer));
 							}
@@ -937,7 +938,7 @@ static void LoadCharacter(const char[] character, int charset, const char[] map,
 							{
 								case KeyValType_Section:
 								{
-									if(!FileExists(key, true))
+									if(check && !FileExists(key, true))
 									{
 										LogError("[Boss] '%s' is missing file '%s' in '%s'", character, key, section);
 									}
@@ -1041,7 +1042,7 @@ static void LoadCharacter(const char[] character, int charset, const char[] map,
 							{
 								case KeyValType_Section:
 								{
-									if(FileExists(key, true))
+									if(!check || FileExists(key, true))
 									{
 										PrecacheModel(key);
 									}
@@ -1054,7 +1055,7 @@ static void LoadCharacter(const char[] character, int charset, const char[] map,
 								{
 									if(length > val.size)	// "models/example.mdl"	"mdl"
 									{
-										if(FileExists(key, true))
+										if(!check || FileExists(key, true))
 										{
 											PrecacheModel(key);
 										}
@@ -1065,7 +1066,7 @@ static void LoadCharacter(const char[] character, int charset, const char[] map,
 									}
 									else if(val.data[0])	// "1"	"models/example.mdl"
 									{
-										if(FileExists(val.data, true))
+										if(!check || FileExists(val.data, true))
 										{
 											PrecacheModel(val.data);
 										}
@@ -1103,7 +1104,7 @@ static void LoadCharacter(const char[] character, int charset, const char[] map,
 							{
 								case KeyValType_Section:
 								{
-									if(FileExists(key, true))
+									if(!check || FileExists(key, true))
 									{
 										AddToStringTable(DownloadTable, key);
 									}
@@ -1121,7 +1122,7 @@ static void LoadCharacter(const char[] character, int charset, const char[] map,
 											for(int b; b < sizeof(MatExts); b++)
 											{
 												FormatEx(buffer, sizeof(buffer), "%s.%s", key, MatExts[b]);
-												if(FileExists(buffer, true))
+												if(!check || FileExists(buffer, true))
 												{
 													AddToStringTable(DownloadTable, buffer);
 												}
@@ -1137,7 +1138,7 @@ static void LoadCharacter(const char[] character, int charset, const char[] map,
 											for(int b; b < sizeof(MdlExts); b++)
 											{
 												FormatEx(buffer, sizeof(buffer), "%s.%s", key, MdlExts[b]);
-												if(FileExists(buffer, true))
+												if(!check || FileExists(buffer, true))
 												{
 													if(b)
 														AddToStringTable(DownloadTable, buffer);
@@ -1151,7 +1152,7 @@ static void LoadCharacter(const char[] character, int charset, const char[] map,
 										}
 										else
 										{
-											if(FileExists(key, true))
+											if(!check || FileExists(key, true))
 											{
 												AddToStringTable(DownloadTable, key);
 											}
@@ -1163,7 +1164,7 @@ static void LoadCharacter(const char[] character, int charset, const char[] map,
 									}
 									else			// "1"	"sound/example.mp3"
 									{
-										if(FileExists(val.data, true))
+										if(!check || FileExists(val.data, true))
 										{
 											AddToStringTable(DownloadTable, val.data);
 										}
@@ -1204,12 +1205,14 @@ static void LoadCharacter(const char[] character, int charset, const char[] map,
 									for(int b; b < sizeof(MdlExts); b++)
 									{
 										FormatEx(buffer, sizeof(buffer), "%s.%s", key, MdlExts[b]);
-										if(FileExists(buffer, true))
+
+										bool phy = b == sizeof(MdlExts)-1;
+										if((!phy && !check) || FileExists(buffer, true))
 										{
 											if(b)
 												AddToStringTable(DownloadTable, buffer);
 										}
-										else if(b != sizeof(MdlExts)-1)
+										else if(!phy)
 										{
 											LogError("[Boss] '%s' is missing file '%s' in '%s'", character, buffer, section);
 											break;
@@ -1223,12 +1226,14 @@ static void LoadCharacter(const char[] character, int charset, const char[] map,
 										for(int b; b < sizeof(MdlExts); b++)
 										{
 											FormatEx(buffer, sizeof(buffer), "%s.%s", key, MdlExts[b]);
-											if(FileExists(buffer, true))
+
+											bool phy = b == sizeof(MdlExts)-1;
+											if((!phy && !check) || FileExists(buffer, true))
 											{
 												if(b)
 													AddToStringTable(DownloadTable, buffer);
 											}
-											else if(b != sizeof(MdlExts)-1)
+											else if(!phy)
 											{
 												LogError("[Boss] '%s' is missing file '%s' in '%s'", character, buffer, section);
 												break;
@@ -1240,12 +1245,14 @@ static void LoadCharacter(const char[] character, int charset, const char[] map,
 										for(int b; b < sizeof(MdlExts); b++)
 										{
 											FormatEx(buffer, sizeof(buffer), "%s.%s", val.data, MdlExts[b]);
-											if(FileExists(buffer, true))
+
+											bool phy = b == sizeof(MdlExts)-1;
+											if((!phy && !check) || FileExists(buffer, true))
 											{
 												if(b)
 													AddToStringTable(DownloadTable, buffer);
 											}
-											else if(b != sizeof(MdlExts)-1)
+											else if(!phy)
 											{
 												LogError("[Boss] '%s' is missing file '%s' in '%s'", character, buffer, section);
 												break;
@@ -1284,7 +1291,7 @@ static void LoadCharacter(const char[] character, int charset, const char[] map,
 									for(int b; b < sizeof(MatExts); b++)
 									{
 										FormatEx(buffer, sizeof(buffer), "%s.%s", key, MatExts[b]);
-										if(FileExists(buffer, true))
+										if(!check || FileExists(buffer, true))
 										{
 											AddToStringTable(DownloadTable, buffer);
 										}
@@ -1301,7 +1308,7 @@ static void LoadCharacter(const char[] character, int charset, const char[] map,
 										for(int b; b < sizeof(MatExts); b++)
 										{
 											FormatEx(buffer, sizeof(buffer), "%s.%s", key, MatExts[b]);
-											if(FileExists(buffer, true))
+											if(!check || FileExists(buffer, true))
 											{
 												AddToStringTable(DownloadTable, buffer);
 											}
@@ -1316,7 +1323,7 @@ static void LoadCharacter(const char[] character, int charset, const char[] map,
 										for(int b; b < sizeof(MatExts); b++)
 										{
 											FormatEx(buffer, sizeof(buffer), "%s.%s", val.data, MatExts[b]);
-											if(FileExists(buffer, true))
+											if(!check || FileExists(buffer, true))
 											{
 												AddToStringTable(DownloadTable, buffer);
 											}
@@ -1645,8 +1652,8 @@ void Bosses_CreateFromConfig(int client, ConfigMap cfg, int team, int leader = 0
 	}
 	
 	Forward_OnBossCreated(client, Client(client).Cfg, !active);
-	
 	Goomba_BossCreated(Client(client).Cfg);
+	Music_BossCreated(client);
 	
 	if(Client(client).Cfg.GetInt("companion", i))
 	{
@@ -3018,59 +3025,64 @@ static void EnableSubplugins()
 {
 	if(!PluginsEnabled)
 	{
-		PluginsEnabled = true;
-		
-		char path[PLATFORM_MAX_PATH], filename[PLATFORM_MAX_PATH], filepath1[PLATFORM_MAX_PATH], filepath2[PLATFORM_MAX_PATH];
-		BuildPath(Path_SM, path, sizeof(path), "plugins/freaks");
-		
-		FileType filetype;
-		DirectoryListing dir = OpenDirectory(path);
-		if(dir)
+		char folder[128];
+		Cvar[SubpluginFolder].GetString(folder, sizeof(folder));
+		if(folder[0])
 		{
-			while(dir.GetNext(filename, sizeof(filename), filetype))
+			PluginsEnabled = true;
+			
+			char path[PLATFORM_MAX_PATH], filename[PLATFORM_MAX_PATH], filepath1[PLATFORM_MAX_PATH], filepath2[PLATFORM_MAX_PATH];
+			BuildPath(Path_SM, path, sizeof(path), "plugins/%s", folder);
+			
+			FileType filetype;
+			DirectoryListing dir = OpenDirectory(path);
+			if(dir)
 			{
-				if(filetype == FileType_File)
+				while(dir.GetNext(filename, sizeof(filename), filetype))
 				{
-					int pos = strlen(filename) - 4;
-					if(pos > 0)
+					if(filetype == FileType_File)
 					{
-						if(StrEqual(filename[pos], ".smx"))
+						int pos = strlen(filename) - 4;
+						if(pos > 0)
 						{
-							FormatEx(filepath1, sizeof(filepath1), "%s/%s", path, filename);
-							
-							if(!IsSubpluginLoaded(filename))
-								InsertServerCommand("sm plugins load freaks/%s", filename);
-							
-							DataPack pack = new DataPack();
-							pack.WriteString(filepath1);
-							RequestFrame(Bosses_RenameSubplugin, pack);
-						}
-						else if(StrEqual(filename[pos], ".ff2"))
-						{
-							FormatEx(filepath1, sizeof(filepath1), "%s/%s", path, filename);
-							
-							strcopy(filename[pos], 5, ".smx");
-							FormatEx(filepath2, sizeof(filepath2), "%s/%s", path, filename);
-							
-							if(FileExists(filepath2))
+							if(StrEqual(filename[pos], ".smx"))
 							{
-								DeleteFile(filepath1);
+								FormatEx(filepath1, sizeof(filepath1), "%s/%s", path, filename);
+								
+								if(!IsSubpluginLoaded(filename))
+									InsertServerCommand("sm plugins load %s/%s", folder, filename);
+								
+								DataPack pack = new DataPack();
+								pack.WriteString(filepath1);
+								RequestFrame(Bosses_RenameSubplugin, pack);
 							}
-							else
+							else if(StrEqual(filename[pos], ".ff2"))
 							{
-								RenameFile(filepath2, filepath1);
-								InsertServerCommand("sm plugins load freaks/%s", filename);
+								FormatEx(filepath1, sizeof(filepath1), "%s/%s", path, filename);
+								
+								strcopy(filename[pos], 5, ".smx");
+								FormatEx(filepath2, sizeof(filepath2), "%s/%s", path, filename);
+								
+								if(FileExists(filepath2))
+								{
+									DeleteFile(filepath1);
+								}
+								else
+								{
+									RenameFile(filepath2, filepath1);
+									InsertServerCommand("sm plugins load %s/%s", folder, filename);
+								}
+								
+								DataPack pack = new DataPack();
+								pack.WriteString(filepath2);
+								RequestFrame(Bosses_RenameSubplugin, pack);
 							}
-							
-							DataPack pack = new DataPack();
-							pack.WriteString(filepath2);
-							RequestFrame(Bosses_RenameSubplugin, pack);
 						}
 					}
 				}
+				
+				ServerExecute();
 			}
-			
-			ServerExecute();
 		}
 	}
 }
@@ -3105,37 +3117,45 @@ public void Bosses_RenameSubplugin(DataPack pack)
 	int pos = strcopy(buffer2, sizeof(buffer2), buffer1) - 4;
 	strcopy(buffer2[pos], 5, ".ff2");
 	
-	RenameFile(buffer2, buffer1);
+	if(!RenameFile(buffer2, buffer1))
+		LogError("Failed to rename '%s' to '%s'", buffer1, buffer2);
 }
 
 static void DisableSubplugins()
 {
 	if(PluginsEnabled)
 	{
-		PluginsEnabled = false;
-		
-		//TODO: Reverse
-		ArrayList list = new ArrayList(PLATFORM_MAX_PATH);
-		
-		char filename[PLATFORM_MAX_PATH];
-		Handle iter = GetPluginIterator();
-		while(MorePlugins(iter))
+		char folder[128];
+		Cvar[SubpluginFolder].GetString(folder, sizeof(folder));
+		if(folder[0])
 		{
-			Handle plugin = ReadPlugin(iter);
-			GetPluginFilename(plugin, filename, sizeof(filename));
-			if(!StrContains(filename, "freaks\\", false))
-				list.PushString(filename);
+			PluginsEnabled = false;
+
+			StrCat(folder, sizeof(folder), "\\");
+
+			ArrayList list = new ArrayList(PLATFORM_MAX_PATH);
+			
+			char filename[PLATFORM_MAX_PATH];
+
+			Handle iter = GetPluginIterator();
+			while(MorePlugins(iter))
+			{
+				Handle plugin = ReadPlugin(iter);
+				GetPluginFilename(plugin, filename, sizeof(filename));
+				if(!StrContains(filename, folder, false))
+					list.PushString(filename);
+			}
+			delete iter;
+			
+			for(int i = list.Length-1; i >= 0; i--)
+			{
+				list.GetString(i, filename, sizeof(filename));
+				InsertServerCommand("sm plugins unload %s", filename);
+			}
+			
+			delete list;
+			ServerExecute();
 		}
-		delete iter;
-		
-		for(int i = list.Length-1; i >= 0; i--)
-		{
-			list.GetString(i, filename, sizeof(filename));
-			InsertServerCommand("sm plugins unload %s", filename);
-		}
-		
-		delete list;
-		ServerExecute();
 	}
 }
